Add analyze pass
This patch moves license conflict detection out of the verify pass and
into the new analyze pass. Now codebom is cleanly separated into
three phases:

1. `scan` for missing declarations
2. `verify` declarations are consistent with the codebase
3. `analyze` the potential implications of the declarations

Fixes #1
