+++ codebom/verify_test.py
@@ -136,57 +136,6 @@ def test_verify_bom_dev_dependencies(httpserver):
     assert verify_bom({'development-dependencies': [{'root': 'foss'}]}) == None
     assert verify_bom({'development-dependencies': [{'root': 'foss', 'license': 'MIT'}]}, is_source_dist=True) == None
 
-def test_collect_license_warnings():
-    def license_warnings(data, is_source_dist=False):
-        return verify.collect_license_warnings(Bom(data, '.'), is_source_dist)
-
-    data = {
-        'license': 'AllRightsReserved',
-        'dependencies': [{'root': 'foss', 'license': 'GPL-3.0'}]
-    }
-    assert license_warnings(data) == ["The license 'AllRightsReserved' may be incompatible with the license 'GPL-3.0' in 'foss'. Specify 'copyright-holders' and/or 'licensees' to state the license is authorized in this context. If this dependency is used only for development, move it to the 'development-dependencies' section."]
-
-    # Declaring AllRightsReserved is not sufficient. Need copyright-holders too.
-    data = {
-        'license': 'AllRightsReserved',
-        'dependencies': [{'root': 'foss', 'license': 'AllRightsReserved'}]
-    }
-    assert license_warnings(data) == ["The license 'AllRightsReserved' may be incompatible with the license 'AllRightsReserved' in 'foss'. Specify 'copyright-holders' and/or 'licensees' to state the license is authorized in this context. If this dependency is used only for development, move it to the 'development-dependencies' section."]
-
-    # Use 'licensee' declaration if you are an authorized client.
-    data = {
-        'license': 'AllRightsReserved',
-        'copyright-holders': ['P', 'Q'],
-        'dependencies': [{'root': 'foss', 'license': 'GPL-3.0', 'licensees': ['Q', 'R']}]
-    }
-    assert license_warnings(data) == []
-
-    # Assume copyright-holder is an implicit licensee.
-    data = {
-        'license': 'AllRightsReserved',
-        'copyright-holders': ['Q'],
-        'dependencies': [{'root': 'foss', 'license': 'GPL-3.0', 'copyright-holders': ['Q']}]
-    }
-    assert license_warnings(data) == []
-
-    # Don't warn if conflict happens within a development dependency (unless
-    # verifying a source distribution).
-    data = {
-        'development-dependencies': [
-            {'license': 'AllRightsReserved', 'dependencies': [{'license': 'GPL-3.0'}]}
-        ]
-    }
-    assert license_warnings(data) == []
-
-    # Don't warn on known conflicts.
-    data = {
-        'license': 'AllRightsReserved',
-        'copyright-holders': ['Q'],
-        'potential-license-conflicts': ['foss'],
-        'dependencies': [{'root': 'foss', 'license': 'GPL-3.0'}]
-    }
-    assert license_warnings(data) == []
-
 def test_warn_if_license_mismatch():
     def license_warnings(data, is_source_dist=False):
         allLicenseIds = ['MIT', 'Apache-1.1', 'Apache-2.0']
@@ -200,8 +149,9 @@ def test_warn_if_license_mismatch():
     assert license_warnings(data) == []
 
 def test_verify_bom_license_conflict(capsys):
-    out = verify_bom({'license': 'AllRightsReserved', 'dependencies': [{'root': 'foss', 'license': 'GPL-3.0'}]})
-    assert out.msg.startswith('The license')
+    licenseFile = os.path.join(os.path.dirname(__file__), 'licenses', 'Apache-2.0.txt')
+    out = verify_bom({'license': 'Apache-1.1', 'license-file': licenseFile})
+    assert out.msg.startswith('License file')
 
 def test_verify_bom_output():
     data = {'licensees': ['Q'], 'potential-license-conflicts': ['foo']}
